Move state into separate file
