+++ modules/internationalisation/src/country.ts
@@ -1,112 +1,5 @@
-export const usStates: Record<string, string> = {
-	AL: 'Alabama',
-	AK: 'Alaska',
-	AZ: 'Arizona',
-	AR: 'Arkansas',
-	AE: 'Armed Forces',
-	AA: 'Armed Forces America',
-	AP: 'Armed Forces Pacific',
-	CA: 'California',
-	CO: 'Colorado',
-	CT: 'Connecticut',
-	DE: 'Delaware',
-	FL: 'Florida',
-	GA: 'Georgia',
-	GU: 'Guam',
-	HI: 'Hawaii',
-	ID: 'Idaho',
-	IL: 'Illinois',
-	IN: 'Indiana',
-	IA: 'Iowa',
-	KS: 'Kansas',
-	KY: 'Kentucky',
-	LA: 'Louisiana',
-	ME: 'Maine',
-	MD: 'Maryland',
-	MA: 'Massachusetts',
-	MI: 'Michigan',
-	MN: 'Minnesota',
-	MS: 'Mississippi',
-	MO: 'Missouri',
-	MT: 'Montana',
-	NE: 'Nebraska',
-	NV: 'Nevada',
-	NH: 'New Hampshire',
-	NJ: 'New Jersey',
-	NM: 'New Mexico',
-	NY: 'New York',
-	NC: 'North Carolina',
-	ND: 'North Dakota',
-	OH: 'Ohio',
-	OK: 'Oklahoma',
-	OR: 'Oregon',
-	PA: 'Pennsylvania',
-	PR: 'Puerto Rico',
-	RI: 'Rhode Island',
-	SC: 'South Carolina',
-	SD: 'South Dakota',
-	TN: 'Tennessee',
-	TX: 'Texas',
-	UT: 'Utah',
-	VT: 'Vermont',
-	VI: 'Virgin Islands',
-	VA: 'Virginia',
-	WA: 'Washington',
-	DC: 'Washington DC (District of Columbia)',
-	WV: 'West Virginia',
-	WI: 'Wisconsin',
-	WY: 'Wyoming',
-};
+import { isInList } from '@modules/arrayFunctions';
 
-export const caStateCodes = [
-	'AB',
-	'BC',
-	'MB',
-	'NB',
-	'NL',
-	'NT',
-	'NS',
-	'NU',
-	'ON',
-	'PE',
-	'QC',
-	'SK',
-	'YT',
-] as const;
-
-export type CaState = (typeof caStateCodes)[number];
-
-export const caStates: Record<CaState, string> = {
-	AB: 'Alberta',
-	BC: 'British Columbia',
-	MB: 'Manitoba',
-	NB: 'New Brunswick',
-	NL: 'Newfoundland and Labrador',
-	NT: 'Northwest Territories',
-	NS: 'Nova Scotia',
-	NU: 'Nunavut',
-	ON: 'Ontario',
-	PE: 'Prince Edward Island',
-	QC: 'Quebec',
-	SK: 'Saskatchewan',
-	YT: 'Yukon',
-} as const;
-
-export const auStates: Record<string, string> = {
-	ACT: 'Australian Capital Territory',
-	NSW: 'New South Wales',
-	NT: 'Northern Territory',
-	QLD: 'Queensland',
-	SA: 'South Australia',
-	TAS: 'Tasmania',
-	VIC: 'Victoria',
-	WA: 'Western Australia',
-};
-export const newspaperCountries: Record<string, string> = {
-	GB: 'United Kingdom',
-	IM: 'Isle of Man',
-};
-export type IsoCountry = (typeof isoCountries)[number];
 export const isoCountries = [
 	'GB',
 	'US',
@@ -359,7 +252,7 @@ export const isoCountries = [
 	'SH',
 ] as const;
 
-export const isoCountrySet: Set<IsoCountry> = new Set(isoCountries);
+export type IsoCountry = (typeof isoCountries)[number];
 
 export const countries: Record<IsoCountry, string> = {
 	GB: 'United Kingdom',
@@ -612,10 +505,18 @@ export const countries: Record<IsoCountry, string> = {
 	JE: 'Jersey',
 	SH: 'Saint Helena',
 };
-export function getCountryNameByIsoCode(code: IsoCountry) {
-	return countries[code] || null;
+
+export const isoCountrySet: Set<IsoCountry> = new Set(isoCountries);
+
+export function isIsoCountry(search: string): search is IsoCountry {
+	return isInList(isoCountries)(search);
 }
-// ----- Types ----- //
-export type UsState = keyof typeof usStates;
-export type AuState = keyof typeof auStates;
-export type StateProvince = UsState;
+
+export const newspaperCountries: Partial<Record<IsoCountry, string>> = {
+	GB: 'United Kingdom',
+	IM: 'Isle of Man',
+};
+
+export function getCountryNameByIsoCode(code: IsoCountry) {
+	return countries[code];
+}
